rename the bn_matrix.py to pswf_utils.py
